Add Manual Advance Mode
This commit adds the ability to enable manual advance mode. This could
be desired for applications that want granular control over when the
trip advances. To enable simply set the delay to `-1`, and advance by
calling `trip.next()`.

Additionally to support manual mode I added prev/next buttons that can
be toggled useing `showNavigation`. The labels are configurable with
`prevLabel` and `nextLabel`.

To give developers more programmatic control I added `canGoPrev` and
`canGoNext` which must return truthy before the trip will proceed.
This can be useful if you want to prevent the user from progressing
until they take an action or to prevent them from going back in the
tour.