If string passed as selector, convert to jQuery object, then restore to string.
The problem with converting the selector into a jQuery object and leaving it as one is if the stored object gets deleted and remade (i.e. an ajax dialog window), the stored object is now incorrect. If a string is passed you can assume you always want the most current existing version of that selector passed.